feat: unwrap sync namespace to flat exports + barrel
